requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
chardet==5.2.0
Django==4.2.5
django-right-to-left==0.1.1
et-xmlfile==1.1.0
jdatetime==4.1.1
Levenshtein==0.23.0
mssql-django==1.3
openpyxl==3.1.2
Pillow==10.0.0
psycopg==3.1.17
psycopg2-binary==2.9.9
pycop==0.0.11
pyodbc==5.0.1
python-barcode==0.15.1
python-bidi==0.4.2
python-Levenshtein==0.23.0
pytz==2023.3.post1
rapidfuzz==3.4.0
reportlab==4.0.8
rtl==0.4.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
validators==0.22.0
wfastcgi==3.0.0
